Fixed Statistics Visualization bugs

+++ GP_template/src/All/Channel.java
@@ -2,8 +2,8 @@ package All;
 
 public class Channel {
 	public int channel;
-	public int probability;
-	public Channel(int channel, int probability){
+	public double probability;
+	public Channel(int channel, double probability){
 		this.channel = channel;
 		this.probability = probability;
 	}
